src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
bottleneck==1.5.0
    # via langchain-google-vertexai
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
docstring-parser==0.17.0
    # via google-cloud-aiplatform
fastapi==0.116.1
    # via nilor-cli-v1 (pyproject.toml)
google-api-core==2.25.1
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   nilor-cli-v1 (pyproject.toml)
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform==1.110.0
    # via langchain-google-vertexai
google-cloud-bigquery==3.36.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.21.0
    # via nilor-cli-v1 (pyproject.toml)
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-storage==2.19.0
    # via
    #   google-cloud-aiplatform
    #   langchain-google-vertexai
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.31.0
    # via google-cloud-aiplatform
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-resource-manager
grpcio==1.74.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.74.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   nilor-cli-v1 (pyproject.toml)
    #   google-genai
    #   langchain-google-vertexai
    #   langgraph-sdk
    #   langsmith
httpx-sse==0.4.1
    # via langchain-google-vertexai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.74
    # via
    #   nilor-cli-v1 (pyproject.toml)
    #   langchain-google-vertexai
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-vertexai==2.0.28
    # via nilor-cli-v1 (pyproject.toml)
langgraph==0.6.6
    # via nilor-cli-v1 (pyproject.toml)
langgraph-checkpoint==2.1.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.6.4
    # via langgraph
langgraph-sdk==0.2.2
    # via langgraph
langsmith==0.4.15
    # via langchain-core
numexpr==2.11.0
    # via langchain-google-vertexai
numpy==2.3.2
    # via
    #   bottleneck
    #   numexpr
    #   shapely
orjson==3.11.2
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   langchain-core
    #   langsmith
    #   pytest
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
protobuf==6.32.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==19.0.1
    # via langchain-google-vertexai
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   nilor-cli-v1 (pyproject.toml)
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
    #   langchain-core
    #   langchain-google-vertexai
    #   langgraph
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via
    #   nilor-cli-v1 (pyproject.toml)
    #   pytest-asyncio
pytest-asyncio==1.1.0
    # via nilor-cli-v1 (pyproject.toml)
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-dotenv==1.1.1
    # via uvicorn
pyyaml==6.0.2
    # via
    #   langchain-core
    #   uvicorn
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
shapely==2.1.1
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
tenacity==9.1.2
    # via
    #   google-genai
    #   langchain-core
typing-extensions==4.14.1
    # via
    #   anyio
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via nilor-cli-v1 (pyproject.toml)
uvloop==0.21.0
    # via uvicorn
validators==0.35.0
    # via langchain-google-vertexai
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
xxhash==3.5.0
    # via langgraph
zstandard==0.24.0
    # via langsmith
